Add parseblock.js
