+++ src/compiler/blocks/modifiers/loop.js
@@ -1,21 +1,8 @@
 export default {
 	name: "loop",
 	description: "iterates over a block of code",
-	execute(amount, block, current) {
-		let anss = []
-		let ans = []
-		let declared = []
-		for (let i = 0; i < amount; i++) block.forEach(e => {
-			if (e.type !== "text") {
-				if (!declared.includes(e)) {
-					ans.push(e)
-				}
-				declared.push(e)
-				return;
-			}
-			else anss.push(e.commands)
-		})
-		anss.push("\n\tret\n")
+	execute(amount, current, anss, ans) {
+		
 		ans.push(
 			{
 				os: ['linux', 'win', 'mac'],
